Removed  Grafana Plugins storage and  dev env

+++ socket/models/speed.js
@@ -2,7 +2,7 @@
 module.exports = (sequelize, DataTypes) => {
   const Speed = sequelize.define('Speed', {
     time: DataTypes.DATE,
-    value: DataTypes.BIGINT,
+    value: DataTypes.FLOAT,
     type: DataTypes.STRING,
   }, {});
   // Speed.associate = function(models) {
